# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CureBery API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 6 */
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  patient: "patient",
  nurse: "nurse",
} as const;

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  role: UserProfileRole;
}

export interface LoginResponse {
  success: boolean;
  user: UserProfile;
  token: string;
}

export interface NursePublicProfile {
  id: number;
  name: string;
  strNumber: string;
  specialization: string;
  isOnline: boolean;
  rating: number;
  distanceKm: number;
  lat: number;
  lng: number;
  avatarUrl?: string;
}

export interface NurseFullProfile {
  id: number;
  name: string;
  email: string;
  strNumber: string;
  specialization: string;
  isOnline: boolean;
  rating: number;
  lat: number;
  lng: number;
  avatarUrl?: string;
  totalPatients: number;
  yearsExperience: number;
}

export interface UpdateLocationRequest {
  lat: number;
  lng: number;
}

export interface UpdateStatusRequest {
  isOnline: boolean;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetNearbyNursesParams = {
  /**
   * Latitude of patient location
   */
  lat: number;
  /**
   * Longitude of patient location
   */
  lng: number;
  /**
   * Radius in kilometers (default 3)
   */
  radius?: number;
};