# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CureBery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from "./userProfile";

export interface LoginResponse {
  success: boolean;
  user: UserProfile;
  token: string;
}