# lib/api-zod/src/generated/types/nurseFullProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CureBery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NurseFullProfile {
  id: number;
  name: string;
  email: string;
  strNumber: string;
  specialization: string;
  isOnline: boolean;
  rating: number;
  lat: number;
  lng: number;
  avatarUrl?: string;
  totalPatients: number;
  yearsExperience: number;
}