# lib/api-zod/src/generated/types/nursePublicProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CureBery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NursePublicProfile {
  id: number;
  name: string;
  strNumber: string;
  specialization: string;
  isOnline: boolean;
  rating: number;
  distanceKm: number;
  lat: number;
  lng: number;
  avatarUrl?: string;
}