# lib/api-zod/src/generated/types/updateLocationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CureBery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLocationRequest {
  lat: number;
  lng: number;
}